Drop runtime dependency on ActiveSupport due to minimum ruby requirement
more_core_extensions/core_ext/object/blank.rb was copied from
https://github.com/rails/rails/blob/v5.0.0/activesupport/lib/active_support/core_ext/object/blank.rb

With the release of ActiveSupport v5.0.0 a minimum required_ruby_version
of '>= 2.2.2' was added.  The code for Blank has been copied here so that
we can continue to support ruby versions that don't meet that requirement.

The test was mostly copied as well, but rewritten in rspec syntax.

ActiveSupport has been added as a development dependency for ruby >= 2.2.2
to allow for testing nested hash features with HashWithIndifferentAccess