Update email regex
Updates the email regex to use that defined in
http://www.regular-expressions.info/email.html. The main reason to do
this is because it is more permissive, and email regexes in general tend
to be overly constrictive. Sure enough, this breaks some of the existing
tests which asserted that certain combinations of characters were
invalid. Importantly, though, no assertions on valid emails were
broken (though the tests here aren't exhaustive, so pinch of salt).

Follow up to https://github.com/ManageIQ/manageiq/issues/12781