Reduce memory usage of `.store_path`
By using recursion to loop through the keys for store path, we are
creating new objects that are not needed every time with go further down
into the stack, because we are required to do a `args[1..-1].push`,
which will create a new array from the subset of the args, and pass it
into the next call of `.store_path`.

By not using recursion, we can completely remove any object allocations
that aren't necessary (besides the ones that are needed for the method's
spec), and still maintain the same functionality.  The added benefit is
that this also now not prone to `stack level too deep` errors because it
no longer requires recursion to function.