Reduce memory usage of `.fetch_path`
By using recursion to loop through the keys for fetch_path, we are
creating new objects that are not needed every time with go further down
into the stack, because we are required to do a `args[1..-1].push`,
which will create a new array from the subset of the args, and pass it
into the next call of `.fetch_path`.

By not using recursion, we can completely remove any object allocations
that aren't necessary (besides the ones that are needed for the method's
spec), and still maintain the same functionality.  The added benefit is
that this also now not prone to `stack level too deep` errors because it
no longer requires recursion to function.

Of note, I attempted to use both `.each_with_object` and `.inject`
instead of implementing something myself.  But each had it's own
problems that were either broken or sub-optimal:

The `.each_with_object` implementation looked something like this:

    args.each.with_index.each_with_object(self) do |(key, i), result|

Obviously not requiring the `result = self` line above it.  This just failed to
function in some scenarios and I was unable to determine the cause.

The `.inject` implementation was similar to the above
`.each_with_object` implementation:

    args.each.with_index.inject(self) do |result, (key, i)|

While actually functioning as expected, and just making sure to end the
block with a final call of `result`, the combination of having to do
`.each`, `.with_index`, and `.inject` to achieve having both a indexed
loop and a resulting object actually caused some extra objects to be
created.

Being explicit here and avoiding the ruby convenience methods seemed to
both prevent extraneous objects instantiation and provide the results
expected.