fix sporadic thread_safety test failure
For this test, the spawned thread was ending after the main block was running

When the main block quit, it undefined the method on the class.
The threaded class kept calling the method (now undefined) and threw an exception
because the method was no longer defined.

solution is to wait for the thread to finish before exiting and undefining the method